# webapitc2005b/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Web API Express",
        "description": "My WebApi using Node",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3001",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/getMiniGame/{minigame_id}": {
            "get": {
                "tags": [
                    "MiniGame"
                ],
                "summary": "Encuentra un minijuego por id",
                "description": "Regresa el minijuego con el id solicitado",
                "operationId": "getMiniGame",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "minigame_id",
                        "description": "Id del minijuego a obtener",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/MiniGame"
                        }
                    }
                }
            }
        },
        "/api/addMiniGame": {
            "post": {
                "tags": [
                    "MiniGame"
                ],
                "summary": "Agrega un nuevo minijuego",
                "description": "Agrega un minijuego a la base de datos",
                "operationId": "addMiniGame",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Nuevo minijuego",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MiniGame"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/MiniGame"
                        }
                    },
                    "405": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/updateMiniGame/{minigame_id}": {
            "put": {
                "tags": [
                    "MiniGame"
                ],
                "summary": "Actualiza un minijuego",
                "description": "Actualiza un minijuegp en la base de datos",
                "operationId": "updateMiniGame",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "minigame_id",
                        "description": "Id del minijuegpo a actualizar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Datos nuevo documento",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MiniGame"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/MiniGame"
                        }
                    },
                    "400": {
                        "description": "Minijuego Inválido"
                    },
                    "404": {
                        "description": "Minijuego no encontrado"
                    }
                }
            }
        },
        "/api/deleteMiniGame/{minigame_id}": {
            "delete": {
                "tags": [
                    "MiniGame"
                ],
                "summary": "Borra un minijuego",
                "description": "Borra un mninijuego de la base de datos",
                "operationId": "deleteMiniGame",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "minigame_id",
                        "description": "Id del mninijuego a borrar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/MiniGame"
                        }
                    },
                    "400": {
                        "description": "Minijuego Inválido"
                    },
                    "404": {
                        "description": "Minijuego no encontrado"
                    }
                }
            }
        },
        "/api/getLevelGroupMiniGame/{level_id}/{group_id}/{minigame_id}": {
            "get": {
                "tags": [
                    "LevelGroupMiniGame"
                ],
                "summary": "Encuentra un minijuego por id",
                "description": "Regresa el minijuego con el id solicitado",
                "operationId": "getMiniGame",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "level_id",
                        "description": "Id del nivel a obtener",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "group_id",
                        "description": "Id del grupo a obtener",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "minigame_id",
                        "description": "Id del minijuego a obtener",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/LevelGroupMiniGame"
                        }
                    }
                }
            }
        },
        "/api/addLevelGroupMiniGame": {
            "post": {
                "tags": [
                    "LevelGroupMiniGame"
                ],
                "summary": "Agrega un nuevo minijuego",
                "description": "Agrega un minijuego a la base de datos",
                "operationId": "addMiniGame",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Nuevo minijuego",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LevelGroupMiniGame"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/LevelGroupMiniGame"
                        }
                    },
                    "405": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/updateLevelGroupMiniGame/{level_id}/{group_id}/{minigame_id}": {
            "put": {
                "tags": [
                    "LevelGroupMiniGame"
                ],
                "summary": "Actualiza un minijuego",
                "description": "Actualiza un minijuegp en la base de datos",
                "operationId": "updateMiniGame",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "level_id",
                        "description": "Id del nivel a obtener",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "group_id",
                        "description": "Id del grupo a obtener",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "minigame_id",
                        "description": "Id del minijuego a obtener",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Datos nuevo documento",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LevelGroupMiniGame"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/LevelGroupMiniGame"
                        }
                    },
                    "400": {
                        "description": "Minijuego Inválido"
                    },
                    "404": {
                        "description": "Minijuego no encontrado"
                    }
                }
            }
        },
        "/api/deleteLevelGroupMiniGame/{levelgroupminigame_id}": {
            "delete": {
                "tags": [
                    "LevelGroupMiniGame"
                ],
                "summary": "Borra un minijuego",
                "description": "Borra un mninijuego de la base de datos",
                "operationId": "deleteMiniGame",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "levelgroupminigame_id",
                        "description": "Id del mninijuego a borrar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/LevelGroupMiniGame"
                        }
                    },
                    "400": {
                        "description": "Minijuego Inválido"
                    },
                    "404": {
                        "description": "Minijuego no encontrado"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "MiniGame": {
            "type": "object",
            "properties": {
                "minigame_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "minigame_name": {
                    "type": "string"
                },
                "minigame_image": {
                    "type": "string"
                }
            }
        },
        "LevelGroupMiniGame": {
            "type": "object",
            "properties": {
                "levelgroupminigame_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "level_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "group_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "minigame_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "document_id": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        }
    }
}